Refactoring for heroku to work
Reference #25
